requirements.txt
torch
transformers
accelerate>=1.0.0
datasets>=3.0.0
peft>=0.13.0
Pillow
pandas
matplotlib
tqdm
PyYAML
qwen-vl-utils
math-verify[antlr4_13_2]
